+++ CHANGELOG.md
@@ -15,6 +15,7 @@ For a full diff see [`0.4.0...main`][0.4.0...main].
 - Allow user to disable daily note processing in plugin settings
 - Allow user to disable book note reading journey processing in plugin settings
 - Remove book and reading journey items when note is deleted
+- Replaced tag pie chart with tag cloud
 
 ### Fixed